backend/requirements.txt
# anyio==3.7.1
# argon2-cffi==21.3.0
# argon2-cffi-bindings==21.2.0
# arrow==1.2.3
asgiref==3.7.2
# asttokens==2.2.1
# async-lru==2.0.4
# attrs==23.1.0
# Babel==2.12.1
# backcall==0.2.0
# beautifulsoup4==4.12.2
# bleach==6.0.0
# cargo==0.3
certifi==2023.7.22
# cffi==1.15.1
charset-normalizer==3.2.0
# colorama==0.4.6
# comm==0.1.4
# cryptography==42.0.5
# debugpy==1.6.7
# decorator==5.1.1
# defusedxml==0.7.1
# distlib==0.3.7
Django==5.0.3
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3
django-mysql==4.12.0
djangorestframework==3.15.0
# djangorestframework-simplejwt==5.3.1
# dnspython==2.4.2
# executing==1.2.0
# Faker==19.10.0
# fastjsonschema==2.18.0
# filelock==3.12.4
# fqdn==1.5.1
idna==3.4
# ipykernel==6.25.1
# ipython==8.14.0
# ipython-genutils==0.2.0
# ipywidgets==8.1.0
# isoduration==20.11.0
# jedi==0.19.0
# Jinja2==3.1.2
# json5==0.9.14
# jsonpointer==2.4
# jsonschema==4.19.0
# jsonschema-specifications==2023.7.1
# jupyter==1.0.0
# jupyter-console==6.6.3
# jupyter-events==0.7.0
# jupyter-lsp==2.2.0
# jupyter_client==8.3.0
# jupyter_core==5.3.1
# jupyter_server==2.7.0
# jupyter_server_terminals==0.4.4
# jupyterlab==4.0.4
# jupyterlab-pygments==0.2.2
# jupyterlab-widgets==3.0.8
# jupyterlab_server==2.24.0
# MarkupSafe==2.1.3
# matplotlib-inline==0.1.6
# mistune==3.0.1
# mysql==0.0.3
# nbclient==0.8.0
# nbconvert==7.7.3
# nbformat==5.9.2
# nest-asyncio==1.5.7
# notebook==7.0.2
# notebook_shim==0.2.3
# numpy==1.26.0
# overrides==7.4.0
# packaging==23.1
# pandocfilters==1.5.0
# parso==0.8.3
# pickleshare==0.7.5
# Pillow==10.0.1
# platformdirs==3.10.0
# prometheus-client==0.17.1
# prompt-toolkit==3.0.39
# protobuf==4.21.12
# psutil==5.9.5
# pure-eval==0.2.2
# pycparser==2.21
# pycryptodome==3.19.0
# Pygments==2.16.1
# PyJWT==2.8.0
# pymongo==4.5.0
PyMySQL==1.1.0
# python-dateutil==2.8.2
# python-json-logger==2.0.7
# pywin32==306
# pywinpty==2.0.11
# PyYAML==6.0.1
# pyzmq==25.1.0
# qtconsole==5.4.3
# QtPy==2.3.1
# razorpay==1.4.1
# referencing==0.30.2
requests==2.31.0
# rfc3339-validator==0.1.4
# rfc3986-validator==0.1.1
# rpds-py==0.9.2
# scipy==1.11.2
# Send2Trash==1.8.2
# six==1.16.0
# sniffio==1.3.0
# sounddevice==0.4.6
# soupsieve==2.4.1
sqlparse==0.4.4
# stack-data==0.6.2
# terminado==0.17.1
# tinycss2==1.2.1
# tornado==6.3.2
# traitlets==5.9.0
tzdata==2023.3
# uri-template==1.3.0
urllib3==2.0.4
# virtualenv==20.24.5
# wavio==0.0.7
# wcwidth==0.2.6
# webcolors==1.13
# webencodings==0.5.1
# websocket-client==1.6.1
# widgetsnbextension==4.0.8
